Track application is attribute
Converted the current running application from a state to an attribute.  Improved the logic for detected changed applications, preventing flip-flop reporting of app from new app, to Roku, and back to new app.  This update also reduces the number of calls to the Roku device from 4 per app launch to only 1.  This change is a pre-cursor to converting the child app from a momentary switch, to a standard switch.