Added new command keyPress()
The new keypress command will enable ability to create custom buttons in RM not yet supported in at the parent device level.  This solution was chosen to simplify the user interface.  A new Navigation child device will be added in the future that has only the features of the Roku Remote (minus the app launchers) to allow for a visual interface to these functions