Custom refresh for current channel / bug fix
Fixed a bug in the scheduler that would cause refreshes to not happen for large periods of time.
Added ability to control the current application refresh on a different schedule than the TV status schedule.
When using separate schedules, the current channel refresh is unscheduled while the TV is off.  If using the TV remote to turn on the TV, and ultimate control channels, then channel changing will not register in hubitat until the tv status is refreshed.